Extract lessons list utilities
